Butterfat of Cows

Random samples of ten mature (five-year-old and older) and ten two-year-old cows were taken from each of five breeds. The average butterfat percentage of these 100 cows is stored in the variable butterfat with the type of cow stored in the variable breed and the age of the cow stored in the variable age.

Format

A data frame with 100 observations on the following 3 variables:

butterfat

average butterfat percentage

age

a factor with levels 2 years old and Mature

breed

a factor with levels Ayrshire, Canadian, Guernsey, Holstein-Friesian, and Jersey

Source

Canadian record book of purebreed diary cattle.

References

Sokal, R. R. and Rohlf, F. J. (1994) Biometry. W. H. Freeman, New York, third edition.

Examples

summary(aov(butterfat ~ breed + age, data = Cows))

#>             Df Sum Sq Mean Sq F value Pr(>F)    
#> breed        4  34.32   8.580  49.909 <2e-16 ***
#> age          1   0.21   0.207   1.204  0.275    
#> Residuals   94  16.16   0.172                   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1