Drosophila Melanogaster

Drosophila contains per diem fecundity (number of eggs laid per female per day for the first 14 days of life) for 25 females from each of three lines of Drosophila melanogaster. The three lines are Nonselected (control), Resistant, and Susceptible. Data are used in Example 11.5.

Format

A data frame with 75 observations on the following 2 variables:

Fecundity

number of eggs laid per female per day for the first 14 days of life

Line

a factor with levels Nonselected, Resistant, and Susceptible

Source

The original measurements are from an experiment conducted by R. R. Sokal (Sokal and Rohlf, 1994, p. 237).

References

Sokal, R. R. and Rohlf, F. J. (1994) Biometry. W. H. Freeman, New York, third edition.

Examples

summary(aov(Fecundity ~ Line, data = Drosophila))

#>             Df Sum Sq Mean Sq F value   Pr(>F)    
#> Line         2   1362   681.1   8.666 0.000424 ***
#> Residuals   72   5659    78.6                     
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>