Cross and Auto Fertilization

Plants' heights in inches obtained from two seeds, one obtained by cross fertilization and the other by auto fertilization, in two opposite but separate locations of a pot are recorded.

Format

A data frame with 15 observations on the following 2 variables:

cross

height of plant in inches

self

height of plant in inches

Source

Darwin, C. (1876) The Effect of Cross and Self-Fertilization in the Vegetable Kingdom

References

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Examples

with(data = Fertilize,
t.test(cross, self))

#> 
#> 	Welch Two Sample t-test
#> 
#> data:  cross and self
#> t = 2.4371, df = 22.164, p-value = 0.02328
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  0.3909566 4.8423767
#> sample estimates:
#> mean of x mean of y 
#>  20.19167  17.57500 
#> 
#>