Grocery Spending

The consumer expenditure survey, created by the U.S. Department of Labor, was administered to 30 households in Watauga County, North Carolina, to see how the cost of living in Watauga county with respect to total dollars spent on groceries compares with other counties. The amount of money each household spent per week on groceries is stored in the variable groceries. Data are used in Example 8.3.

Format

A data frame with 30 observations on the following variable:

groceries

total dollars spent on groceries

Source

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Examples

with(data = Grocery,
z.test(x = groceries, sigma.x = 30, conf.level = 0.97)$conf)

#> [1] 108.7473 132.5194
#> attr(,"conf.level")
#> [1] 0.97
#>