High School Wrestlers

The body fat of 78 high school wrestlers was measured using three separate techniques, and the results are stored in the data frame HSwrestler. The techniques used were hydrostatic weighing (HWFAT), skin fold measurements (SKFAT), and the Tanita body fat scale (TANFAT). Data are used in Examples 10.11, 12.11, and 12.12.

Format

A data frame with 78 observations on the following 9 variables:

AGE

age of wrestler in years

HT

height of wrestler in inches

WT

weight ofwrestler in pounds

ABS

abdominal fat

TRICEPS

tricep fat

SUBSCAP

subscapular fat

HWFAT

hydrostatic fat

TANFAT

Tanita fat

SKFAT

skin fat

Source

Data provided by Dr. Alan Utter, Department of Health Leisure and Exercise Science, Appalachian State University.

References

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Examples

FAT <- c(HSwrestler$HWFAT, HSwrestler$TANFAT, HSwrestler$SKFAT)
GROUP <- factor(rep(c("HWFAT", "TANFAT", "SKFAT"), rep(78, 3)))
BLOCK <- factor(rep(1:78, 3))
friedman.test(FAT ~ GROUP | BLOCK)

#> 
#> 	Friedman rank sum test
#> 
#> data:  FAT and GROUP and BLOCK
#> Friedman chi-squared = 22.487, df = 2, p-value = 1.309e-05
#>