Sign Test

This function will test a hypothesis based on the sign test and reports linearly interpolated confidence intervals for one sample problems.

SIGN.test(x, y = NULL, md = 0, alternative = "two.sided",
  conf.level = 0.95)

Arguments

x

numeric vector; NAs and Infs are allowed but will be removed.

y

optional numeric vector; NAs and Infs are allowed but will be removed.

md

a single number representing the value of the population median specified by the null hypothesis

alternative

is a character string, one of "greater", "less", or "two.sided", or the initial letter of each, indicating the specification of the alternative hypothesis. For one-sample tests, alternative refers to the true median of the parent population in relation to the hypothesized value of the median.

conf.level

confidence level for the returned confidence interval, restricted to lie between zero and one

Value

A list of class htest, containing the following components:

Details

Computes a “Dependent-samples Sign-Test” if both x and y are provided. If only x is provided, computes the “Sign-Test”.

Note

The reported confidence interval is based on linear interpolation. The lower and upper confidence levels are exact.

Null Hypothesis

For the one-sample sign-test, the null hypothesis is that the median of the population from which x is drawn is md. For the two-sample dependent case, the null hypothesis is that the median for the differences of the populations from which x and y are drawn is md. The alternative hypothesis indicates the direction of divergence of the population median for x from md (i.e., "greater", "less", "two.sided".)

References

Gibbons, J.D. and Chakraborti, S. (1992). Nonparametric Statistical Inference. Marcel Dekker Inc., New York.

Kitchens, L.J.(2003). Basic Statistics and Data Analysis. Duxbury.

Conover, W. J. (1980). Practical Nonparametric Statistics, 2nd ed. Wiley, New York.

Lehmann, E. L. (1975). Nonparametrics: Statistical Methods Based on Ranks. Holden and Day, San Francisco.

See also

z.test, zsum.test, tsum.test

Examples

with(data = Phone,
SIGN.test(call.time,md=2.1))

#> 
#> 	One-sample Sign-Test
#> 
#> data:  call.time
#> s = 11, p-value = 1
#> alternative hypothesis: true median is not equal to 2.1
#> 95 percent confidence interval:
#>  0.7261939 5.4952244
#> sample estimates:
#> median of x 
#>         2.1 
#> 
#> 

#>                   Conf.Level L.E.pt U.E.pt
#> Lower Achieved CI     0.9069 0.8000 5.2000
#> Interpolated CI       0.9500 0.7262 5.4952
#> Upper Achieved CI     0.9653 0.7000 5.6000
#> 

        # Example 10.1 from PASWR.
        # Computes two-sided sign-test for the null hypothesis
        # that the population median is 2.1.  The alternative
        # hypothesis is that the median is not 2.1.  An interpolated
        # upper 95% upper bound for the population median will be computed.

Author
Alan T. Arnholt