World Cup Soccer

Soccer contains how many goals were scored in the regulation 90 minute periods of World Cup soccer matches from 1990 to 2002. Data are used in Example 4.4.

Format

A data frame with 575 observations on the following 3 variables:

CGT

cumulative goal time in minutes

Game

game in which goals were scored

Goals

number of goals scored in regulation period

Source

Chu, S. (2003) “Using Soccer Goals to Motivate the Poisson Process.” INFORMS Transaction on Education, 3, 2: 62-68.

Details

The World Cup is played once every four years. National teams from all over the world compete. In 2002 and in 1998, thirty-six teams were invited; whereas, in 1994 and in 1990, only 24 teams participated. The data frame Soccer contains three columns: CGT, Game, and Goals. All of the information contained in Soccer is indirectly available from the FIFA World Cup website, located at http://fifaworldcup.yahoo.com/.

References

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Examples

with(data = Soccer,
table(Goals))

#> Goals
#>  0  1  2  3  4  5  6  7  8 
#> 19 49 60 47 32 18  3  3  1