Tire Stopping Distances

The data frame Tire has the stopping distances measured to the nearest foot for a standard sized car to come to a complete stop from a speed of sixty miles per hour. There are six measurements of the stopping distance for four different tread patterns labeled A, B, C, and D. The same driver and car were used for all twenty-four measurements. Data are used in Example 11.1 and 11.2.

Format

A data frame with 24 observations on the following 2 variables:

StopDist

stopping distance measured to the nearest foot

tire

a factor with levels A, B, C, and D

Source

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Examples

summary(aov(StopDist ~ tire, data = Tire))

#>             Df Sum Sq Mean Sq F value  Pr(>F)   
#> tire         3   5673  1891.0   5.328 0.00732 **
#> Residuals   20   7099   354.9                   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>