Weight Gain in Rats

The data come from an experiment to study the gain in weight of rats fed on four different diets, distinguished by amount of protein (low and high) and by source of pretein (beef and cereal).

Format

A data frame with 40 observations on the following 4 variables.

ProteinSource

a factor with levels Beef Cereal

ProteinAmount

a factor with levels High Low

weightgain

weight gain in grams

Source

D. J. Hand, F. Daly, A. D. Lunn, K. J. McConway and E. Ostrowski (1994) A Handbook of Small Datasets. Chapman and Hall/CRC, London.

Details

The design of the experiment is acompletely randomized with ten rats on each of the four treatments.

Examples

aov(weightgain ~ ProteinSource*ProteinAmount, data = WeightGain)

#> Call:
#>    aov(formula = weightgain ~ ProteinSource * ProteinAmount, data = WeightGain)
#> 
#> Terms:
#>                 ProteinSource ProteinAmount ProteinSource:ProteinAmount
#> Sum of Squares          220.9        1299.6                       883.6
#> Deg. of Freedom             1             1                           1
#>                 Residuals
#> Sum of Squares     8049.4
#> Deg. of Freedom        36
#> 
#> Residual standard error: 14.95307
#> Estimated effects may be unbalanced
#>