Apartments in Vitoria

Descriptive information and the appraised total price (in Euros) for apartments in Vitoria, Spain.

Format

A data frame with 218 observations on the following 16 variables:

row.labels

the number of the observation

totalprice

the market total price (in Euros) of the apartment including garage(s) and storage room(s)

area

the total living area of the apartment in square meters

zone

a factor indicating the neighborhood where the apartment is located with levels Z11, Z21, Z31, Z32, Z34, Z35, Z36, Z37, Z38, Z41, Z42, Z43, Z44, Z45, Z46, Z47, Z48, Z49, Z52, Z53, Z56, Z61, and Z62.

category

a factor indicating the condition of the apartment with levels 2A, 2B, 3A, 3B, 4A, 4B, and 5A. The factors are ordered so that 2A is the best and 5A is the worst.

age

age of the aprtment

floor

floor on which the apartment is located

rooms

total number of rooms including bedrooms, dining room, and kitchen

out

a factor indicating the percent of the apartment exposed to the elements. The levels E100, E75, E50, and E25, correspond to complete exposure, 75% exposure, 50% exposure, and 25% exposure respectively.

conservation

is an ordered factor indicating the state of conservation of the apartment. The levels 1A, 2A, 2B, and 3A are ordered from best to worst conservation.

toilets

the number of bathrooms

garage

the number of garages

elevator

indicates the absence (0) or presence (1) of elevators.

streetcategory

an ordered factor from best to worst indicating the category of the street with levels S2, S3, S4, and S5

heating

a factor indicating the type of heating with levels 1A, 3A, 3B, and 4A which correspond to: no heating, low-standard private heating, high-standard private heating, and central heating respectively.

tras

the number of storage rooms outside of the apartment

Source

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Examples

modTotal <- lm(totalprice ~ area + as.factor(elevator) +
area:as.factor(elevator), data = vit2005)
modSimpl <- lm(totalprice ~ area, data = vit2005)
anova(modSimpl,modTotal)

#> Analysis of Variance Table
#> 
#> Model 1: totalprice ~ area
#> Model 2: totalprice ~ area + as.factor(elevator) + area:as.factor(elevator)
#>   Res.Df        RSS Df Sum of Sq      F    Pr(>F)    
#> 1    216 3.5970e+11                                  
#> 2    214 3.0267e+11  2 5.704e+10 20.165 9.478e-09 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

rm(modSimpl, modTotal)