Swim Times

Swimmers' improvements in seconds for two diets are stored in the data frame Swimtimes. The values in highfat represent the time improvement in seconds for swimmers on a high fat diet, and the values in lowfat represent the time improvement in seconds for swimmers on a low fat diet. Data are used in Example 10.9.

Format

A data frame with 14 observations on the following 2 variables:

lowfat

time improvement in seconds

highfat

time improvement in seconds

Source

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. (2008) Probability and Statistics with R. Chapman & Hall/CRC.

Details

Times for the thirty-two swimmers for the 200 yard individual medley were taken right after the swimmers' conference meet. The swimmers were randomly assigned to follow one of the diets. The group on diet 1 followed a low fat diet the entire year but lost two swimmers along the way. The group on diet 2 followed the high fat diet the entire year and also lost two swimmers.

Examples

with(data = Swimtimes,
wilcox.test(highfat, lowfat))

Warning message:
cannot compute exact p-value with ties

#> 
#> 	Wilcoxon rank sum test with continuity correction
#> 
#> data:  highfat and lowfat
#> W = 143, p-value = 0.04072
#> alternative hypothesis: true location shift is not equal to 0
#> 
#>